Calculate Hours of Temperature Stress
Source: R/calculate_stress_hours.R

Calculate Hours of Temperature Stress

calculate_hours(interpolated_surfaces, location, threshold)

Arguments

location
A given location for which the temperature threshold is to be checked against. Should be provided as an points or polygon object.

threshold
A numeric value against which interpolated temperatures are checked for exceeding.

forecast_surface
A list of interpolated hourly temperature surfaces from interpolate_hourly.

Details

Calculates hours temperature stress given a list of interpolated surfaces of a BoM six-day forecast from interpolate_hourly.

Examples

# NOT RUN {
library(sp)
locations <-
 structure(
   list(
     longitude = c(
       151.9134,
       151.7387
     ),
     latitude = c(
       -27.7952,
       -27.5425
     )
   ),
   .Names = c("longitude",
              "latitude"),
   class = "data.frame",
   row.names = c(NA, -2L))

xy <- locations[, c(1, 2)]

locations <- SpatialPointsDataFrame(
 coords = xy,
 data = location,
 proj4string = CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
)

mungbean_stress <- calculate_hours(x, location = locations, 30)
# }