Test BoM Precis Forecast for Changes and Send an Alert on Changes or Failure
Source: R/unit_test_forecast.R

Test BoM Precis Forecast for Changes and Send an Alert on Changes or Failure

unit_test_forecast(AUS_forecast, project_email_address, project_email_password,
  report_email_addresses, yesterdays_forecast)

Arguments

AUS_forecast
A forecast object generated by get_precis_forecast to check for changes.

project_email_address
Project e-mail account used to send alerts.

project_email_password
Password for e-mail account used to send alerts.

report_email_addresses
An atomic vector of e-mail addresses to which error reports should be sent.

yesterdays_forecast
A csv file of the prior day's forecast used as basis for checking current day's forecast for changes

Value

E-mail if issues are encountered.

Details

Tests the current BoM forecast against the prior day's forecast for changes and sends an e-mail alert on change or failure.

Examples

# NOT RUN {
unit_test(forecast = AUS_forecast)
# }